pyproject.toml
[build-system]
requires=["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "teyvat-lore-graph"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.0",
    "flask>=3.1.3",
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.9",
    "langchain-google-genai",
    "langchain-groq>=1.1.2",
    "langchain-neo4j>=0.8.0",
    "langchain-ollama>=1.0.1",
    "neo4j",
    "pydantic",
    "python-dotenv",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.2",
    "streamlit",
]

[tool.setuptools.packages.find]
where = ["."]
